refactor: Refactor role management: unify hooks, etc
