--- src/api/endpoints/roles.ts
@@ -1,6 +1,0 @@
-const roleEndpoints = {
-  getAllRoles: '/role',
-  createRole: '/role/create-role',
-};
-
-export default roleEndpoints;
