+++ src/api/services/index.ts
@@ -1,4 +1,3 @@
 export * as specialtyService from './specialty';
 export * as authService from './authService';
 export * as userService from './userService';
-export * as roleService from './role';
