--- src/features/role/mutation/useCreateRole.ts
@@ -1,11 +1,0 @@
-import { roleService } from '@api/services';
-import { useMutation } from '@tanstack/react-query';
-import type { ICreateRole } from 'src/modules/shared/types/role.type';
-
-const useCreateRoleQuery = () => {
-  return useMutation({
-    mutationFn: async (role: ICreateRole) => roleService.createRole(role),
-  });
-};
-
-export default useCreateRoleQuery;
