+++ src/modules/role/hooks/useCreateRole.ts
@@ -4,14 +4,17 @@ import { useTheme } from '@mui/material';
 import { useEffect, useState } from 'react';
 
 import useGetAllPermissionGrouped from '@modules/permissions/hooks/useGetAllPermissionGrouped';
-import useCreateRoleQuery from '@features/role/mutation/useCreateRole';
-import useNotificationStore from 'src/modules/shared/stores/useNotificationStore';
 
-import useGetOneRoleQuery from '@features/role/query/useGetOneRoleQuery';
-import type { IBasicIdNameDescription } from 'src/modules/shared/types/common.types';
+import type { IBasicIdNameDescription } from '@type/common.types';
+import useNotificationStore from '@stores/useNotificationStore';
+
+import {
+  useUpdateRole,
+  useCreateRoleQuery,
+  useGetOneRoleQuery,
+} from './useRolQueries';
 import { INITIAL_VALUES } from '../constants';
 import type { IFormValues } from '../types';
-import useUpdateRole from '@features/role/mutation/useUpdateRole';
 
 function useCreateRole() {
   const { id } = useParams();
