--- src/page/Role/ListRols/types/index.ts
@@ -1,5 +1,0 @@
-export interface IFormValues {
-  name: string;
-  description: string;
-  permission_id: number[];
-}
