--- src/page/Role/index.ts
@@ -1,2 +1,0 @@
-export { default as CreateRole } from './CreateRole';
-export { default as ListRols } from './ListRols';
